requirements.txt
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
matplotlib==3.3.1
numpy==1.19.1
# opendp-smartnoise==0.1.3.1
opendp-smartnoise==0.1.4
pandas==1.1.0
scipy==1.5.2
seaborn==0.11.0
z3-solver==4.8.9.0

# for notebook testing
#nbclient==0.5.3
